requirements.txt
chalice==1.2.3
pyyaml==3.12
click==6.6
